Refactor to separate Vonage class
