+++ tests/Feature/ClientPrivateKeyCredentialsTest.php
@@ -9,7 +9,7 @@ class ClientPrivateKeyCredentialsTest extends FeatureTestCase
 {
     protected function getEnvironmentSetUp($app)
     {
-        $app['config']->set('vonage.private_key', '/path/to/key');
+        $app['config']->set('vonage.private_key', __DIR__.'/../fixtures/private.key');
         $app['config']->set('vonage.application_id', 'application-id-123');
     }
