+++ tests/fixtures/private.key
@@ -1,0 +1,1 @@
+===FAKE-KEY===
